Change view of some elements, add sure to delete restaurant, stylizeng some lements

+++ src/main/java/dwr/company/restauracje/popupWindowsControler.java
@@ -40,7 +40,7 @@ public class popupWindowsControler implements Initializable{
     private Spinner amountToAdd;
     private final ObservableList<Storage> storageList = FXCollections.observableArrayList();
     private Storage selectedIngridient;
-
+    private boolean yesORno;
     private final ObservableList<Storage> componentProductList = FXCollections.observableArrayList();
 
     private final List <Storage> ingridients = new ArrayList<>();
@@ -78,6 +78,8 @@ public class popupWindowsControler implements Initializable{
     private Button deleteCategory;
     @FXML
     private Button yesButton, noButton;
+    @FXML
+    private Button yesButtonR, noButtonR;
     @FXML
     private ComboBox newPlace;
     @FXML
@@ -128,11 +130,15 @@ public class popupWindowsControler implements Initializable{
             case 3:
                 break;
             case 4:
+                yesORno = false;
                 try {
                     comboWithRestaurants.getItems().addAll(Client.getRestaurantNames());
                 } catch (IOException e) {
                     e.printStackTrace();
                 }
+                break;
+            case 5:
+                yesORno = false;
                 break;
             default:
                 break;
@@ -286,8 +292,23 @@ public class popupWindowsControler implements Initializable{
     }
 
     @FXML
+    protected void warningWindow2() throws IOException {
+        GeneralController.popup = 5;
+        FXMLLoader loader = new FXMLLoader(App.class.getResource("deleteRestaurant.fxml"));
+        Scene popupScene = new Scene(loader.load(), 437.0, 167.0);
+        //popupScene.getStylesheets().add(getClass().getResource("style.css").toExternalForm());
+        popupScene.getStylesheets().add(Objects.requireNonNull(getClass().getResource("style.css")).toExternalForm());
+        Stage popup = new Stage();
+        popup.initStyle(StageStyle.UNDECORATED);
+        popup.setScene(popupScene);
+        moveWindow(popupScene, popup);
+        popup.show();
+    }
+
+    @FXML
     protected void decision(ActionEvent event){
         if(event.getTarget().equals(yesButton)){
+            yesORno = true;
             Stage stage = (Stage) yesButton.getScene().getWindow();
             stage.close();
         }else if (event.getTarget().equals(noButton)){
@@ -431,14 +452,15 @@ public class popupWindowsControler implements Initializable{
     }
 
     public void deleteRestaurant() throws IOException {
-        if(comboWithRestaurants.getSelectionModel().isEmpty()){
-            restaurantsWaringLabel.setText("wybierz restauracje");
-        }
-        else{
-            //delete
-            Client.deleteRestaurant(comboWithRestaurants.getValue().toString());
-            comboWithRestaurants.getItems().clear();
-            comboWithRestaurants.getItems().addAll(Client.getRestaurantNames());
+        if(yesORno){
+            if (comboWithRestaurants.getSelectionModel().isEmpty()) {
+                restaurantsWaringLabel.setText("wybierz restauracje");
+            } else {
+                //delete
+                Client.deleteRestaurant(comboWithRestaurants.getValue().toString());
+                comboWithRestaurants.getItems().clear();
+                comboWithRestaurants.getItems().addAll(Client.getRestaurantNames());
+            }
         }
     }
